% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amortisationschedule.R
\name{amortisationschedule}
\alias{amortisationschedule}
\title{Title amortisationschedule}
\usage{
amortisationschedule(amount, monthlyinterestrate, paymentsperyear, years)
}
\arguments{
\item{amount}{principal}

\item{monthlyinterestrate}{monthly interest rate}

\item{paymentsperyear}{period}

\item{years}{years}
}
\value{
\preformatted{                A list with repayment schedule and basic loan information.
}
}
\description{
amortisationschedule
}
\examples{
df <- amortisationschedule(1000, 0.03, 3, 1)

df

}
