% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMT.R
\name{IPMT}
\alias{IPMT}
\title{Title IPMT}
\usage{
IPMT(rate, per, nper, pv, fv = 0, type = 0)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Interest portion of instalment in each period
}
\examples{
IPMT(0.36/12, 1, 12*1, 1000)
}
