% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr}
\alias{irr}
\title{Title irr}
\usage{
irr(x, period = 1, starting.value = 0.1)
}
\arguments{
\item{starting.value}{}
}
\value{

}
\description{
Takes a vector of payments and returns a list which includes the internal rate of return ($IRR) and possible word of warning ($beware)
}
\examples{
irr(c(1000,-353.5304,-353.5304,-353.5304))$IRR * 12
}
