% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMT.R
\name{PPMT}
\alias{PPMT}
\title{Title PPMT}
\usage{
PPMT(rate, per, nper, pv, fv = 0, type = 0)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Principal portion of instalment in each period
}
\examples{
PPMT(0.36/12, 1, 12*1, 1000)
}
