% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMT.R
\name{IPMT}
\alias{IPMT}
\title{Title IPMT}
\usage{
IPMT(rate, per, nper, pv, fv = 0, type = 0)
}
\arguments{
\item{rate}{Required. Lending rates.}

\item{per}{Required. The number of periods used to calculate its interest amount, must be between 1 and nper.}

\item{nper}{Required. The total number of payments for this loan.}

\item{pv}{Required. Present value, or the sum of what a series of future payments are worth now, is also called principal.}

\item{fv}{Optional. The future value, or the expected cash balance after the last payment. If fv is omitted, its value is assumed to be 0 (zero), i.e. the future value of the loan is 0.}

\item{type}{Optional. The number 0 (zero) or 1 indicates the payout time.}
}
\value{

}
\description{
Returns the amount of interest repayment on an investment for a given number of periods, based on a fixed interest rate and equal installments.
}
\examples{
IPMT(0.36/12, 1, 12*1, 1000)
}
