% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMT.R
\name{PPMT}
\alias{PPMT}
\title{Title PPMT}
\usage{
PPMT(rate, per, nper, pv, fv = 0, type = 0)
}
\arguments{
\item{rate}{Required. Lending rates.}

\item{per}{Required. The number of periods used to calculate its interest amount, must be between 1 and nper.}

\item{nper}{Required. The total number of payments for this loan.}

\item{pv}{Required. Present value, or the sum of what a series of future payments are worth now, is also called principal.}

\item{fv}{Optional. The future value, or the expected cash balance after the last payment. If fv is omitted, its value is assumed to be 0 (zero), i.e. the future value of the loan is 0.}

\item{type}{Optional. The number 0 (zero) or 1 indicates the payout time.}
}
\value{

}
\description{
Returns the principal payment over a known period of an investment based on regular fixed payments and a fixed interest rate.
}
\examples{
PPMT(0.36/12, 1, 12*1, 1000)
}
